more updates

+++ src/s-z/StochRsi/StochRsi.Models.cs
@@ -3,6 +3,6 @@ namespace Skender.Stock.Indicators;
 [Serializable]
 public class StochRsiResult : ResultBase
 {
-    public decimal? StochRsi { get; set; }
-    public decimal? Signal { get; set; }
+    public double? StochRsi { get; set; }
+    public double? Signal { get; set; }
 }
